LPS-33427 SF - move from DDMStructure to JournalArticle
