+++ portal-impl/src/com/liferay/portal/service/permission/SubscriptionPermissionImpl.java
@@ -31,8 +31,6 @@ import com.liferay.portlet.bookmarks.model.BookmarksEntry;
 import com.liferay.portlet.bookmarks.service.permission.BookmarksEntryPermission;
 import com.liferay.portlet.documentlibrary.model.DLFileEntry;
 import com.liferay.portlet.documentlibrary.service.permission.DLFileEntryPermission;
-import com.liferay.portlet.dynamicdatamapping.model.DDMStructure;
-import com.liferay.portlet.dynamicdatamapping.service.permission.DDMStructurePermission;
 import com.liferay.portlet.journal.model.JournalArticle;
 import com.liferay.portlet.journal.service.permission.JournalArticlePermission;
 import com.liferay.portlet.messageboards.model.MBCategory;
@@ -168,16 +166,6 @@ public class SubscriptionPermissionImpl implements SubscriptionPermission {
 		else if (className.equals(DLFileEntry.class.getName())) {
 			return DLFileEntryPermission.contains(
 				permissionChecker, classPK, actionId);
-		}
-		else if (className.equals(DDMStructure.class.getName())) {
-			Group group = GroupLocalServiceUtil.fetchGroup(classPK);
-
-			if (group == null) {
-				return DDMStructurePermission.contains(
-					permissionChecker, classPK, actionId);
-			}
-
-			return true;
 		}
 		else if (className.equals(JournalArticle.class.getName())) {
 			return JournalArticlePermission.contains(
