+++ portal-impl/src/com/liferay/portlet/dynamicdatamapping/service/impl/DDMStructureLocalServiceImpl.java
@@ -1257,45 +1257,6 @@ public class DDMStructureLocalServiceImpl
 	}
 
 	/**
-	 * Subscribe the user to changes in elements that belongs to specified
-	 * structure.
-	 *
-	 * @param  groupId the primary key of the folder's group
-	 * @param  userId the primary key of the user to be subscribed
-	 * @param structureId the primary key of the structure to subscribe to
-	 * @throws PortalException if the user, group os structure could not be
-	 * 		   found, or if subscribing was not permissible
-	 * @throws SystemException if a system exception occurred
-	 */
-	@Override
-	public void subscribeStructure(long groupId, long userId, long structureId)
-		throws PortalException, SystemException {
-
-		getSubscriptionLocalService().addSubscription(
-			userId, groupId, DDMStructure.class.getName(), structureId);
-	}
-
-	/**
-	 * Unsubscribe the user from changes in elements that belongs to specified
-	 * structure.
-	 *
-	 * @param  groupId the primary key of the folder's group
-	 * @param  userId the primary key of the user to be subscribed
-	 * @param structureId the primary key of the structure to subscribe to
-	 * @throws PortalException if the user, group os structure could not be
-	 * 		   found, or if subscribing was not permissible
-	 * @throws SystemException if a system exception occurred
-	 */
-	@Override
-	public void unsubscribeStructure(
-		long groupId, long userId, long structureId) throws PortalException,
-		SystemException {
-
-		getSubscriptionLocalService().deleteSubscription(
-			userId, DDMStructure.class.getName(), structureId);
-	}
-
-	/**
 	 * Updates the structure matching the class name ID, structure key, and
 	 * group, replacing its old parent structure, name map, description map, and
 	 * XSD with new ones.
