+++ portal-impl/src/com/liferay/portlet/journal/service/impl/JournalArticleLocalServiceImpl.java
@@ -4409,6 +4409,46 @@ public class JournalArticleLocalServiceImpl
 	}
 
 	/**
+	 * Subscribe the user to changes in elements that belongs to specified
+	 * structure.
+	 *
+	 * @param  groupId the primary key of the folder's group
+	 * @param  userId the primary key of the user to be subscribed
+	 * @param ddmStructureId the primary key of the structure to subscribe to
+	 * @throws PortalException if the user, group os structure could not be
+	 * 		   found, or if subscribing was not permissible
+	 * @throws SystemException if a system exception occurred
+	 */
+	@Override
+	public void subscribeStructure(
+			long groupId, long userId, long ddmStructureId)
+		throws PortalException, SystemException {
+
+		subscriptionLocalService.addSubscription(
+			userId, groupId, DDMStructure.class.getName(), ddmStructureId);
+	}
+
+	/**
+	 * Unsubscribe the user from changes in elements that belongs to specified
+	 * structure.
+	 *
+	 * @param  groupId the primary key of the folder's group
+	 * @param  userId the primary key of the user to be subscribed
+	 * @param ddmStructureId the primary key of the structure to subscribe to
+	 * @throws PortalException if the user, group os structure could not be
+	 * 		   found, or if subscribing was not permissible
+	 * @throws SystemException if a system exception occurred
+	 */
+	@Override
+	public void unsubscribeStructure(
+		long groupId, long userId, long ddmStructureId) throws PortalException,
+		SystemException {
+
+		subscriptionLocalService.deleteSubscription(
+			userId, DDMStructure.class.getName(), ddmStructureId);
+	}
+
+	/**
 	 * Updates the web content article matching the version, replacing its
 	 * folder, title, description, content, and layout UUID.
 	 *
